Fix rdairplay password and rdpanel segfault
Removes the INSTANCE column from RDAIRPLAY and RDPANEL tables.  This was removed in schema 219 but persists if you create a db from scratch.

This makes rdairplay prompt for a password when exiting and seg faults rdpanel if not removed due to an attempt to insert a record into RDAIRPLAY/RDPANEL without setting a value for INSTANCE which is NOT NULL with no default.